chore: remove compiler warning
The compiler emits a warning about officeHours() because its state mutability
can be restricted to pure. However, we want functions overriding it to be able
to read state, and that is why it has a view modifier. So in order to remove
this misplaced warning, the function now reads a state variable. That variable
is not set constant because the warning would reappear.